refactor: 💡 move Stats class into a separate file
